Move classes NodeMapping and NodeMappings to util package
